Duck

+++ src/main/java/com/qlyshopphone_backend/dto/CartDTO.java
@@ -1,4 +1,5 @@
 package com.qlyshopphone_backend.dto;
+import static com.qlyshopphone_backend.constant.ErrorMessage.*;
 
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
@@ -7,12 +8,12 @@ import java.math.BigDecimal;
 
 @Data
 public class CartDTO {
-    private Integer cartId;
-    private Integer productId;
+    private Long cartId;
+    private Long productId;
     private String productName;
     private BigDecimal price;
-    private Integer userId;
-    private Integer quantity;
+    private Long userId;
+    private Long quantity;
     private boolean sold;
     private boolean deleteCart;
     private MultipartFile file;
