+++ src/main/java/com/qlyshopphone_backend/dto/CustomerInfoDTO.java
@@ -1,13 +1,14 @@
 package com.qlyshopphone_backend.dto;
+import static com.qlyshopphone_backend.constant.ErrorMessage.*;
 
 import lombok.Data;
 
 @Data
 public class CustomerInfoDTO {
-    private Integer customerId;
+    private Long customerId;
     private String customerName;
     private String phone;
     private String address;
     private boolean deleteCustomerInfo;
-    private Integer userId;
+    private Long userId;
 }
