+++ src/main/java/com/qlyshopphone_backend/dto/NotificationDTO.java
@@ -1,4 +1,5 @@
 package com.qlyshopphone_backend.dto;
+import static com.qlyshopphone_backend.constant.ErrorMessage.*;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
@@ -7,9 +8,9 @@ import java.time.LocalDateTime;
 
 @Data
 public class NotificationDTO {
-    private Integer notificationId;
+    private Long notificationId;
     private String message;
-    private Integer userId;
+    private Long userId;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime timestamp;
 }
